Addition of flexible body.content container
The current body.content container has a fixed height (set by min-height in CSS).  Changes implemented in CSS and the steps.js file to set the height of the container based on the content inside, and set the min-height style when  using a vertical step layout to prevent the container from becoming smaller/shorter than the steps.